chore(tests): disable signal tests on MacOS
